every model moved out of models/models.py, env.py fixed
