+++ app/server/api/credit_transfer_api.py
@@ -3,10 +3,11 @@ from flask.views import MethodView
 from sqlalchemy import or_
 
 from server import db
-from server.models.models import Token
+from server.models.token import Token
 from server.models.utils import paginate_query
-from server.models.credit_transfer import CreditTransfer, BlockchainTransaction
-from server.models.transfer_account import BlockchainAddress
+from server.models.credit_transfer import CreditTransfer
+from server.models.blockchain_transaction import BlockchainTransaction
+from server.models.blockchain_address import BlockchainAddress
 from server.schemas import credit_transfers_schema, credit_transfer_schema, view_credit_transfers_schema
 from server.utils.auth import requires_auth
 from server.utils.access_control import AccessControl
