+++ app/server/models/blockchain_transaction.py
@@ -1,0 +1,24 @@
+from server import db
+from server.models.utils import ModelBase
+
+
+class BlockchainTransaction(ModelBase):
+    __tablename__ = 'blockchain_transaction'
+
+    status = db.Column(db.String)  # PENDING, SUCCESS, FAILED
+    message = db.Column(db.String)
+    block = db.Column(db.Integer)
+    submitted_date = db.Column(db.DateTime)
+    added_date = db.Column(db.DateTime)
+    hash = db.Column(db.String)
+    nonce = db.Column(db.Integer)
+    transaction_type = db.Column(db.String)
+
+    is_bitcoin = db.Column(db.Boolean)
+
+    # Output spent txn for bitcoin
+    has_output_txn = db.Column(db.Boolean, default=False)
+
+    credit_transfer_id = db.Column(db.Integer, db.ForeignKey("credit_transfer.id"))
+
+    signing_blockchain_address_id = db.Column(db.Integer, db.ForeignKey('blockchain_address.id'))
