+++ app/server/models/chatbot_state.py
@@ -1,0 +1,19 @@
+import datetime
+
+from server import db
+from server.models.utils import ModelBase
+
+
+class ChatbotState(ModelBase):
+    __tablename__ = 'chatbot_state'
+
+    transfer_initialised = db.Column(db.Boolean, default=False)
+    target_user_id = db.Column(db.Integer, default=None)
+    transfer_amount = db.Column(db.Integer, default=0)
+    prev_pin_failures = db.Column(db.Integer, default=0)
+    last_accessed = db.Column(db.DateTime, default=datetime.datetime.utcnow)
+
+    provider_message_id = db.Column(db.String())
+
+    user = db.relationship(
+        'User', backref='chatbot_state', lazy=True, uselist=False)
