+++ app/server/models/feedback.py
@@ -2,6 +2,7 @@ from server import db
 from server.models.utils import ModelBase
 from server.models.user import User
 
+
 class Feedback(ModelBase):
     __tablename__ = 'feedback'
