+++ app/server/models/saved_filter.py
@@ -1,0 +1,11 @@
+from sqlalchemy.dialects.postgresql import JSON
+
+from server import db
+from server.models.utils import ModelBase
+
+
+class SavedFilter(ModelBase):
+    __tablename__ = 'saved_filter'
+
+    name = db.Column(db.String)
+    filter = db.Column(JSON)
