+++ app/server/models/transfer_account.py
@@ -1,29 +1,19 @@
 import datetime
 from sqlalchemy.sql import func
-import os, base64
-from eth_utils import keccak
 from flask import current_app
-from cryptography.fernet import Fernet
 from sqlalchemy.ext.hybrid import hybrid_property
-from web3 import Web3
-
-from server.utils.misc import hex_private_key_to_address
-from server.constants import (
-    ALLOWED_BLOCKCHAIN_ADDRESS_TYPES
-)
 from server import db
 from server.models.utils import ModelBase, OneOrgBase, user_transfer_account_association_table
-from server.models.models import Token
+from server.models.token import Token
 from server.models.user import User
-from server.models.credit_transfer import CreditTransfer, BlockchainTransaction
-from server.utils.blockchain_tasks import (
-    send_eth,
-    make_approval
-)
+from server.models.spend_approval import SpendApproval
+from server.models.credit_transfer import CreditTransfer
+from server.models.blockchain_transaction import BlockchainTransaction
 from server.utils.blockchain_tasks import (
     create_blockchain_wallet
 )
 from server.utils.transfer_enums import TransferStatusEnum
+
 
 class TransferAccount(OneOrgBase, ModelBase):
     __tablename__ = 'transfer_account'
@@ -194,95 +184,3 @@ class TransferAccount(OneOrgBase, ModelBase):
         if organisation:
             self.organisation = organisation
             self.token = organisation.token
-
-class SpendApproval(ModelBase):
-    __tablename__ = 'spend_approval'
-
-    eth_send_task_id = db.Column(db.Integer)
-    approval_task_id = db.Column(db.Integer)
-    receiving_address = db.Column(db.String)
-
-    token_id                      = db.Column(db.Integer, db.ForeignKey(Token.id))
-    giving_transfer_account_id    = db.Column(db.Integer, db.ForeignKey(TransferAccount.id))
-
-    def __init__(self, transfer_account_giving_approval, address_getting_approved):
-
-        self.giving_transfer_account = transfer_account_giving_approval
-
-        self.token = transfer_account_giving_approval.token
-
-        self.receiving_address = address_getting_approved
-
-        eth_send_task_id = send_eth(signing_address=address_getting_approved,
-                                    recipient_address=transfer_account_giving_approval.blockchain_address,
-                                    amount_wei=0.00184196 * 10**18)
-
-        approval_task_id = make_approval(signing_address=transfer_account_giving_approval.blockchain_address,
-                                         token=self.token,
-                                         spender=address_getting_approved,
-                                         amount=1000000,
-                                         dependent_on_tasks=[eth_send_task_id])
-
-        self.eth_send_task_id = eth_send_task_id
-        self.approval_task_id = approval_task_id
-
-class BlockchainAddress(OneOrgBase, ModelBase):
-    __tablename__ = 'blockchain_address'
-
-    address             = db.Column(db.String())
-    encoded_private_key = db.Column(db.String())
-
-    # Either "MASTER", "TRANSFER_ACCOUNT" or "EXTERNAL"
-    type = db.Column(db.String())
-
-    transfer_account_id = db.Column(db.Integer, db.ForeignKey(TransferAccount.id))
-
-    signed_transactions = db.relationship('BlockchainTransaction',
-                                          backref='signing_blockchain_address',
-                                          lazy='dynamic',
-                                          foreign_keys='BlockchainTransaction.signing_blockchain_address_id')
-
-    credit_sends = db.relationship('CreditTransfer', backref='sender_blockchain_address',
-                                   lazy='dynamic', foreign_keys='CreditTransfer.sender_blockchain_address_id')
-
-    credit_receives = db.relationship('CreditTransfer', backref='recipient_blockchain_address',
-                                      lazy='dynamic', foreign_keys='CreditTransfer.recipient_blockchain_address_id')
-
-    @hybrid_property
-    def decrypted_private_key(self):
-
-        fernet_encryption_key = base64.b64encode(keccak(text=current_app.config['SECRET_KEY']))
-        cipher_suite = Fernet(fernet_encryption_key)
-
-        return cipher_suite.decrypt(self.encoded_private_key.encode('utf-8')).decode('utf-8')
-
-    def encrypt_private_key(self, unencoded_private_key):
-
-        fernet_encryption_key = base64.b64encode(keccak(text=current_app.config['SECRET_KEY']))
-        cipher_suite = Fernet(fernet_encryption_key)
-
-        return cipher_suite.encrypt(unencoded_private_key.encode('utf-8')).decode('utf-8')
-
-    def calculate_address(self, private_key):
-        self.address = hex_private_key_to_address(private_key)
-
-    def allowed_types(self):
-        return ALLOWED_BLOCKCHAIN_ADDRESS_TYPES
-
-    def __init__(self, type, blockchain_address=None):
-
-        if type not in self.allowed_types():
-            raise Exception("type {} not one of {}".format(type, self.allowed_types()))
-
-        self.type = type
-
-        if blockchain_address:
-            self.address = blockchain_address
-
-        if self.type == "TRANSFER_ACCOUNT" and not blockchain_address:
-
-            hex_private_key = Web3.toHex(keccak(os.urandom(4096)))
-
-            self.encoded_private_key = self.encrypt_private_key(hex_private_key)
-
-            self.calculate_address(hex_private_key)
