+++ app/server/utils/chatbot_controller.py
@@ -1,7 +1,7 @@
 import datetime, re
 from flask import current_app
 from server import db
-from server.models.models import ChatbotState
+from server.models.chatbot_state import ChatbotState
 from server.models.user import User
 from server.utils.phone import proccess_phone_number
 from server.utils.credit_transfers import make_payment_transfer
